libpwq: prevent pthread_cond_wait/pthread_cond_signal race
It's possible for the following sequence of events to happen:

Worker Thread			Manager Thread
-------------------------------------------------------------
Call wqlist_scan while          Work is queued to the wq.
holding wqlist_mtx, find
nothing.

                                pthread_cond_signal

pthread_cond_wait

That would cause the worker to miss the signal and sleep
indefinitely. Ensure that that can't happen by holding the
mutex when signaling the condition variable.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>